requirements.txt
google-cloud-aiplatform
vertexai
python-dotenv
replicate
python-dotenv
streamlit